+++ CMakeLists.txt
@@ -1,5 +1,9 @@
-cmake_minimum_required(3.29)
+cmake_minimum_required(VERSION 3.29)
 
 project(Propuesta-Final-Examen)
 
-add_executable(Propuesta-Final-Examen Pitagorica.cpp)
+add_executable(Propuesta-Final-Examen
+        Funciones.cpp
+        Funciones.h
+        main.cpp
+)
