+++ main.cpp
@@ -1,0 +1,6 @@
+#include "Funciones.h"
+
+int main() {
+    calcular_triples_pitagoricos();
+    return 0;
+}
